Fold testplot_common into testplot_args
testplot_common.py was imported only by testplot_args.py; merge the
shared helpers (plot-list loader, JSON validation, run_haiku) into the
single CLI script and drop the extra module.
