+++ CLAUDE.md
@@ -75,7 +75,7 @@ pytest -k pattern                               # by name
 Layout:
 - `tests/unit/` — per-module: `test_calculate.py`, `test_phases.py`, `test_poly.py`, `test_refine.py`, `test_resample.py`, `test_softplus.py`, `test_whitney.py`; subdirs `interpolate/` (`test_g_calphad.py`, `test_polyfit.py`, `test_redlich_kister.py`, `test_sgte.py`, `test_softplus_fit.py`, `test_stitched_fit.py`), `phases/` (`test_ase.py`, `test_vectorize.py`), `plot/` (`test_1d_plots.py`, `test_axis.py`, `test_colors.py`, `test_excess_free_energy.py`, `test_polygons.py`).
 - `tests/regression/` — `test_issue_1.py`, `test_issue_51.py`, `test_issue_155.py`, `test_pandas_groupby.py`, `test_single_group_apply.py`, `test_cluster_phase_pandas3.py`.
-- `tests/integration/testplots.py` — **not a pytest test**; a render script that writes phase-diagram PNGs to `tests/integration/_plots/`. Triggered by the `testplot` label (`.github/workflows/testplot.yml`; Haiku also auto-selects a plot subset from changed files — PR #152) or by `@testplot ...` comments (`testplot-mention.yml`, parser uses Haiku to map free-text to `--only`/`--poly-method`/`--tielines`/`diff` flags against an allow-list frozen on `main`; see PR #149 for the parse/render/publish security split, PR #150 for diff-against-main rendering). The Haiku call sites live in `.github/scripts/testplot_args.py` (subcommands `parse` for mentions and `select` for the label workflow); shared helpers and the allow-list loader are in `.github/scripts/testplot_common.py`. The Haiku CLI runs as a pure text→JSON transform (PR #212: `--system-prompt replace`, `--tools ""`, isolated cwd) and the prompt's plot list is built from `testplots.py` docstrings (`PLOT_LIST` in `testplot_common.py`) so keys never go stale. `run_haiku` retries once and falls back to "render everything" on a final JSON parse failure. Images are pushed to the `testplots` gallery branch and inlined into a PR comment.
+- `tests/integration/testplots.py` — **not a pytest test**; a render script that writes phase-diagram PNGs to `tests/integration/_plots/`. Triggered by the `testplot` label (`.github/workflows/testplot.yml`; Haiku also auto-selects a plot subset from changed files — PR #152) or by `@testplot ...` comments (`testplot-mention.yml`, parser uses Haiku to map free-text to `--only`/`--poly-method`/`--tielines`/`diff` flags against an allow-list frozen on `main`; see PR #149 for the parse/render/publish security split, PR #150 for diff-against-main rendering). The Haiku call sites and shared helpers (plot-list loader, JSON validation, `run_haiku`) all live in `.github/scripts/testplot_args.py` (subcommands `parse` for mentions and `select` for the label workflow). The Haiku CLI runs as a pure text→JSON transform (PR #212: `--system-prompt replace`, `--tools ""`, isolated cwd) and the prompt's plot list is built from `testplots.py` docstrings (`PLOT_LIST`) so keys never go stale. `run_haiku` retries once and falls back to "render everything" on a final JSON parse failure. Images are pushed to the `testplots` gallery branch and inlined into a PR comment.
 - `tests/conftest.py` provides shared `two_phase_ideal` / `three_phase_regular_solution` phase fixtures.
 - Uses **Hypothesis** for property-based tests. Optional-dep tests use a module-level `pytestmark = pytest.mark.skipif(ImportAlarm(...).message is not None, ...)` when a whole file depends on one extra (see `tests/unit/phases/test_ase.py`); files mixing independent extras use per-test `@pytest.mark.skipif` against `try/except`-set flags instead (see `tests/unit/test_poly.py`). Do not use `try/except ImportError` as the gate by itself (PR #111).
 
@@ -161,7 +161,7 @@ Cheat sheet of what's been considered. Fetch the issue before re-litigating.
 - Refiner strategy (PR #115), `_set_axis_for` extraction (PR #122), `_join_phase_unit` DRY (PR #132).
 - `boundary_id` int column tagging refined rows by coexistence line (PR #163, closed #124). `_CCBase.run()` assigns one id per `_pair_key`, base `Refiner.run()` one id per non-empty `solve()` call. `RefinedPoint`/`RefinedMiscibilityGap` carry `boundary_id: int = 0`.
 - Plain symmetric buffer trimming landed (PR #127); both the boundary-Voronoi alternative (PR #129) and apex-cleanup follow-up (PR #130) were **rejected**.
-- `testplot` label + `@testplot` mention workflow (PR #103, #148, #149, #150, #152). PR-controlled code is sandboxed away from repo write secrets; label workflow Haiku-picks the plot subset from changed files (#152); mention parser supports `diff` against main (#150). Parse/select are in `.github/scripts/testplot_args.py` with shared helpers in `testplot_common.py`. Haiku robustness rebuilt in two passes: PR #181 added an AST-based plot-name reader + retry/JSON-fallback, and PR #212 finally closed #180 by running the Haiku CLI as a pure text→JSON transform (`--system-prompt replace`, `--tools ""`, isolated cwd) and deriving the prompt's plot list from `testplots.py` docstrings so keys can never drift.
+- `testplot` label + `@testplot` mention workflow (PR #103, #148, #149, #150, #152). PR-controlled code is sandboxed away from repo write secrets; label workflow Haiku-picks the plot subset from changed files (#152); mention parser supports `diff` against main (#150). Parse/select and their shared helpers are all in `.github/scripts/testplot_args.py`. Haiku robustness rebuilt in two passes: PR #181 added an AST-based plot-name reader + retry/JSON-fallback, and PR #212 finally closed #180 by running the Haiku CLI as a pure text→JSON transform (`--system-prompt replace`, `--tools ""`, isolated cwd) and deriving the prompt's plot list from `testplots.py` docstrings so keys can never drift.
 - 1d phase-diagram segmentation by stability flips, not distance clustering (PR #179, closed #177). `_assign_segment_ids(df, scan_col)` is threshold-free and grid-density independent; `_auto_threshold` / `_max_normed_gap` removed.
 - 1d plot legend rewritten as top-spine ticks + right-end phase labels (PR #190, closed #188), with `reference_phase` y-axis subtraction (PR #189, closed #187) and `top_labels`/`side_labels` toggles (both default `True`). `plot_excess_free_energy` legend now appears when only line phases are stable (PR #191, closed #182); `col_wrap` clamped to len(temperatures) (PR #186, closed #183).
 - `plot_excess_free_energy` `f_excess` reference now picks the most stable endpoint phase per group (PR #166, closed #155); per-phase reference selection with 1% concentration-span tolerance.
